cli-node: refactor role utils into PackageRoles + document, move stuff back
