+++ packages/cli-node/src/index.ts
@@ -21,3 +21,4 @@
  */
 
 export * from './git';
+export * from './roles';
