+++ packages/cli-node/src/monorepo/PackageGraph.ts
@@ -17,7 +17,7 @@
 import path from 'path';
 import { getPackages, Package } from '@manypkg/get-packages';
 import { paths } from '../paths';
-import { PackageRole } from '../role';
+import { PackageRole } from '../roles';
 import { GitUtils } from '../git';
 import { Lockfile } from './Lockfile';
 import { JsonValue } from '@backstage/types';
